+++ src/components/coupon/CouponCheckItem.tsx
@@ -1,9 +1,9 @@
 import styled from "@emotion/styled";
-import { useCouponContext } from "../../../pages/order-confirm/context/CouponProvider";
 import CheckboxLabel from "../inputs/CheckboxLabel";
 import Flex from "../styled/Flex";
 import LabelTextPair, { labelTextPairType } from "./LabelTextPair";
-import { Coupon } from "../../../api/coupon";
+import { useCouponContext } from "../../pages/order-confirm/context/CouponProvider";
+import { Coupon } from "../../api/coupon";
 
 interface CouponCheckItemProps {
   coupon: Coupon;
